requirements.txt
python-dotenv==1.1.1
websockets==15.0.1
